requirements.txt
pytest>=7.0.0
playwright>=1.30.0
allure-pytest>=2.9.45